requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.5.0
opencv-python>=4.5.5
torch>=1.10.0
torchvision>=0.11.0
tqdm>=4.62.0
Pillow>=8.3.0
PyYAML>=6.0
ultralytics>=8.0.0  # For YOLOv8
scikit-learn>=1.0.0
albumentations>=1.1.0 